CAST-23087: Option to use homeDirectory as source of username in passwd
A customer has a need to preserve the case of the username in the
passwd file generated by keycloak-localization. The username field
from Keycloak doesn't preserve the case of the username but instead
lowercases it. They were able to work around this in 1.2 by
modifying the Ansible script that exported the users to use the
homeDirectory field instead, which does preserve case.

In 1.3 the Ansible code was converted to python and the customer
needs a way to have case-preserving usernames. To provide this
function, the keycloak_localize script is enhanced to take a new
environment variable USER_EXPORT_NAME_SOURCE which if set to
'homeDirectory' uses the homeDirectory field from the user
attributes rather than the username. If the env var is anything
other than 'homeDirectory' the username field is used as before.

The cray-keycloak-users-localize chart is changed to set the
USER_EXPORT_NAME_SOURCE env var from the user_export_name_source
field in the keycloak-localize ConfigMap which is created by
the keycloak-users playbook during localization. It's optional
and by default the env var isn't set and the username field is
used, as before.

(cherry picked from commit 0f9a15927fa9fd7526fd95b236a3d61f3426322d)