CASMPET-3675: setup tools refresh token when OAuth2Error
After the changes to switch away from Ansible there was an error
where keycloak-users-localize was stuck in a loop where the
exception raised was
`oauthlib.oauth2.rfc6749.errors.InvalidGrantError: (invalid_grant)
Token is not active`.

InvalidGrantError is an instance of OAuth2Error and I think it's
safer to just reset the token any time one of these errors is
encountered.

I also made the change to keycloak-setup in case it ever happens
there.