CASMPET-4394: Configure Keycloak JVM to configure container support
A customer saw a Keycloak instance fail due to running out of heap
memory. The default heap memory configuration is

"-Xms64m -Xmx512m -XX:MetaspaceSize=96M -XX:MaxMetaspaceSize=256m"

In the next minor version of the codecentric keycloak chart after
this one the default is:

"-XX:+UseContainerSupport -XX:MaxRAMPercentage=50.0"

The newer chart also defaults to the Keycloak image that
we're already using.

So the intent is to use newer settings and avoid running out of
heap memory.

I also increased the default memory limit. This is so that when
Keycloak is restarted automatically there's more memory space
so that it's less likely to fail when restarting.