CASMPET-4644: keycloak-users-localize exits on unrecoverable failure
There have been a few cases where keycloak-users-localize gets stuck in
a loop where it's reporting that a group doesn't exist. This is usually
caused by what appears to be the LDAP synchronization either not
happening or only partially finishing. We're often not able to look at
what the original problem was because the Job has been restarted during
the install and the old pods are gone, or the old Job has been looping
so long that the logs have rolled over.

The proposed fix is to have keycloak-users-localize fail and exit
rather than loop in these cases where we've seen the system not
recover. This also requires changing the Job/Pod to not restart the
container when it exits with a failure because otherwise K8s just
restarts the container forever. Also it should be necessary to re-try
the Pod 6 times so the backoffLimit now defaults to 3 attempts.

During my testing I saw that when the Pod fails the first time it
creates the LDAP integration and then the next time it runs the LDAP
integration is already there so the next time it works. Since we want
the Job to show that it failed I added code to clean up the integration
so that it will get recreated the next time and either fail the same
way (or succeed if the problem fixed itself somehow).