CASMPET-4692: Change default user to host's nobody
On our systems the nobody user on the host is

```
nobody:x:65534:65534:nobody:/var/lib/nobody:/bin/bash
```

This sets the USER for the keycloak-setup image to this same UID
and GID, rather than using the default which is the root user.
The goal is to improve security of the system.