CASMPET-5158: Bump versions to prep for 1.2 release
This release includes the following changes:

* CASMPET-4299: Create a Nexus client
* CASMPET-4714: Create OAuth2-Proxy Client

Incremented the MAJOR number for the cray-keycloak-setup image;
the change added a new option which is required so this change
was not backwards-compatible.

Incremented the MAJOR number for the cray-keycloak chart; the
new option is also required for the chart to pass through to the
image.

Incremented the MINOR number for the cray-keycloak-users-localize
chart, this is the first release for 1.2 and it's just picking up
the new cray-keycloak-setup image.