CASMPET-5159: Stop creating keycloak-gatekeeper client
The keycloak-gatekeeper client is no longer being used since the
keycloak-gatekeeper-ingress is no longer being deployed, so the
keycloak-setup tool no longer needs to create it.

The configuration options for the name of the client and secret
are left around because the keycloak-setup tool may be enhanced
to delete the client and Secret.