CASMPET-5155: keycloak-setup creates oauth2-proxy clients for bi-can
The bifurcated CAN feature requires changes to the way that the
keycloak-setup tool is configuring the client for the oauth2-proxy.
There will be a separate oauth2-proxy instance for each of the
networks and they will serve different hosts, so each one requires a
different client be created with the right hosts.

Rather than use the previous method of having env vars for each
setting, there's a new env var that contains a JSON string with the
settings for the different clients. This is easier to code up than the
separate env vars.

The cray-keycloak chart is enhanced to pass in this new env var and
also configures the JSON string.

The oauth2-proxy client is no longer created since it's no longer used.