CASMPET-4277: Bump versions to prep for 1.2 release
This release includes the following change:

* CASMPET-5155: keycloak-setup creates oauth2-proxy clients for bi-can

The image changed, it no longer creates the oauth2-proxy client, so
this is a major number increase.

The cray-keycloak chart changed, it's got new options for the that
must be overridden, so this is a major number increase.

The cray-keycloak-users-localize chart is just picking up the new image
and this chart has been released in 1.2 before so that's a patch number
increase.