CASMPET-5171: Bump versions to prep for release
This release will include the following change:

* CASMPET-5171: keycloak-setup stop creating gatekeeper client

Since this removes the gatekeeper client, I bumped the MAJOR
number on the keycloak-setup image.

The cray-keycloak Helm chart changes to pick up that image and also
a Role had to change, so I bumped the MINOR number.

The cray-keycloak-users-localize chart changes to pick up the image
so just the PATCH number.